Moved _is_memory_dataset
